Moving Handler and HandlingFeatures to a separate file.
