+++ src/index.ts
@@ -1,8 +1,9 @@
 export { attachSockets } from "./attach";
 export { createConfig, type Config } from "./config";
 export { ActionsFactory } from "./actions-factory";
-export type { ActionMap, HandlingFeatures } from "./action";
+export type { ActionMap } from "./action";
 
 // issue 952
 export type { EmissionMap } from "./emission";
 export type { LoggerOverrides } from "./logger";
+export { HandlingFeatures } from "./handler";
